Add Logout button to sidebar
A 'Logout' button has been added to the sidebar in the dashboard UI. The related CSS styles have also been updated to provide better visual design and interaction. The button utilizes the 'MdLogout' icon from 'react-icons/md' library. It is styled to look identical to the other menu items. However, we're using a button because later we will implement a server action via a form action.