Add Pre-functional Search and Pagination components to UsersPage
Introduces a dedicated Search component and a Pagination component to the UsersPage. The components are just a styled representation of the final components, but are not yet functional in any way. The search box in the UsersPage has been replaced with the new Search component while the commented Pagination component's placeholder has been swapped out for the actual component.