Add debounce to search input
The search functionality has been enhanced by adding a debouncing mechanism, which includes a 300ms delay before the search query is submitted. This delay improves efficiency as it reduces the number of unnecessary calls per second when the user is typing. The implementation also required the addition of 'use-debounce' package.